unified setup script

+++ setup.sh
@@ -1,6 +1,7 @@
 #!/bin/bash 
 
 export machine=${HOSTNAME}
+export CAFPYANA_DIR=$(pwd)
 
 # try: get SAM Web from FNAL tools env. Do this first, since next
 # spack will override this one
@@ -30,27 +31,29 @@ spack load xrootd@5.6.9%gcc@12.5.0 arch=linux-almalinux9-x86_64_v2
 #### setup virtual python env if it is not already set
 ######################################################
 # Define the Python version and virtual environment name
-cd envs
-LOGDIR=logs
+ENV_DIR=${CAFPYANA_DIR}/envs
+LOGDIR=${ENV_DIR}/logs
 mkdir -p ${LOGDIR}
+
 # Define the Python version and virtual environment name
 PYTHON_VERSION=3.9.21
 VENV_NAME=venv_py39_cafpyana
+VENV_DIR="${ENV_DIR}/${VENV_NAME}"
 
 # Check if virtual environment already exists
-if [ -d "$VENV_NAME" ]; then
+if [ -d "${VENV_DIR}" ]; then
     echo "Virtual environment '$VENV_NAME' already exists. Activating it."
 else
     # Create the virtual environment
-    python -m venv $VENV_NAME
+    python -m venv ${VENV_DIR}
     echo "Virtual environment '$VENV_NAME' created."
 fi
 
 # Activate the virtual environment
-source $VENV_NAME/bin/activate
+source ${VENV_DIR}/bin/activate
 
 # check requirements are installed
-if pip freeze -r pip_requirements.txt 2>&1 | grep -q "not installed"; then
+if pip freeze -r ${VENV_DIR}/pip_requirements.txt 2>&1 | grep -q "not installed"; then
     echo "Updating virtual env installation"
 
     PIPLOG=${LOGDIR}/init_pip.log
@@ -64,7 +67,6 @@ fi
 echo "Virtual environment '$VENV_NAME' set up successfully with Python $PYTHON_VERSION and required packages installed."
 echo "If you want to exit from this virtual env, do $ deactivate"
 echo "If you want to activate this virtual evn again, do $ source '$VENV_NAME'/bin/activate"
-cd ..
 
 ######################################################
 
@@ -83,20 +85,26 @@ cd ..
 # need to install uuid in the EAF
 ######################################################
 if [[ $machine == *jupyter* ]]; then
-    cd envs
-    echo "Installing uuid for since you are in EAF"
-    wget https://www.kernel.org/pub/linux/utils/util-linux/v2.39/util-linux-2.39.3.tar.xz
-    tar xvf util-linux-2.39.3.tar.xz
-    rm util-linux-2.39.3.tar.xz
-    cd util-linux-2.39.3
-    ./configure --prefix="$(pwd)/local" --disable-all-programs --enable-libuuid
-    make -j$(nproc)
-    make install
-    export C_INCLUDE_PATH="$(pwd)/local/include:$C_INCLUDE_PATH"
-    export CPLUS_INCLUDE_PATH="$(pwd)/local/include:$CPLUS_INCLUDE_PATH"
-    export LD_LIBRARY_PATH="$(pwd)/local/lib:$LD_LIBRARY_PATH"
-    export PKG_CONFIG_PATH="$(pwd)/local/lib/pkgconfig:$PKG_CONFIG_PATH"
-    cd ../
+    export C_INCLUDE_PATH="${ENV_DIR}/local/include:$C_INCLUDE_PATH"
+    export CPLUS_INCLUDE_PATH="${ENV_DIR}/local/include:$CPLUS_INCLUDE_PATH"
+    export LD_LIBRARY_PATH="${ENV_DIR}/local/lib:$LD_LIBRARY_PATH"
+    export PKG_CONFIG_PATH="${ENV_DIR}/local/lib/pkgconfig:$PKG_CONFIG_PATH"
+
+    # check for UUID
+    if [ ! -f "${ENV_DIR}/local/lib/libuuid.so" ]; then
+       cd ${ENV_DIR}
+       echo "Installing uuid for since you are in EAF"
+       UUIDLOG=${LOGDIR}/init_uuid.log
+       echo $(date) >> ${UUIDLOG}
+       wget https://www.kernel.org/pub/linux/utils/util-linux/v2.39/util-linux-2.39.3.tar.xz
+       tar xvf util-linux-2.39.3.tar.xz
+       rm util-linux-2.39.3.tar.xz
+       cd util-linux-2.39.3
+       ./configure --prefix="${ENV_DIR}/local" --disable-all-programs --enable-libuuid | tee -a ${UUIDLOG}
+       make -j$(nproc) | tee -a ${UUIDLOG}
+       make install | tee -a ${UUIDLOG}
+       cd ${CAFPYANA_DIR}
+    fi
 fi
 
 ######################################################
@@ -104,25 +112,25 @@ fi
 ###################################################### 
 python -c "import XRootD" > /dev/null 2>&1 || {
     echo "Could not import XRootD! Attempting to install from source..."
-    cd envs
+    cd ${ENV_DIR}
     XROOTLOG=${LOGDIR}/init_xroot.log
     echo $(date) >> ${XROOTLOG}
 
     OLDPATH=$PATH
-    PATH=$PATH:$PWD
+    PATH=$PATH:${ENV_DIR}
     ln -s /cvmfs/larsoft.opensciencegrid.org/products/cmake/v3_22_2/Linux64bit+3.10-2.17/bin/cmake cmake3
     echo "Using cmake at $(which cmake3)" | tee -a ${XROOTLOG}
     wget https://files.pythonhosted.org/packages/fd/4f/419b8caec575ab4133f41c37c39bd742251a4dc6a208a97a3fd772031fe7/xrootd-5.6.9.tar.gz
     tar -zxvf xrootd-5.6.9.tar.gz
     rm xrootd-5.6.9.tar.gz
     cd xrootd-5.6.9
-    python setup.py install 2>&1 | tee -a ../${XROOTLOG}
-    cd ../..
+    python setup.py install 2>&1 | tee -a ${XROOTLOG}
+    cd ${CAFPYANA_DIR}
     PATH=$OLDPATH
 }
 
 ###################################################### 
 
-export PYTHONPATH=$PYTHONPATH:$PWD/..
-export LD_LIBRARY_PATH=$VIRTUAL_ENV/lib/python3.9/site-packages/xrootd-5.6.9-py3.9-linux-x86_64.egg/pyxrootd:$LD_LIBRARY_PATH
-export CAFPYANA_WD=`pwd`
+export PYTHONPATH=$PYTHONPATH:${CAFPYANA_DIR}
+export LD_LIBRARY_PATH=${VENV_DIR}/lib/python3.9/site-packages/xrootd-5.6.9-py3.9-linux-x86_64.egg/pyxrootd:$LD_LIBRARY_PATH
+export CAFPYANA_WD=${CAFPYANA_DIR}
